requirements.txt
pandas
numpy
scikit-learn
scipy
matplotlib
maxminddb-geolite2
seaborn